Change has_visited to uint8_t
Since it is a boolean value we don't need all of that storage.

+++ nav.h
@@ -12,7 +12,7 @@ extern "C"
 struct nav_cell
 {
     int flood_num;
-    int has_visited;
+    uint8_t has_visited;
     uint8_t wall;
     int row;
     int column;
